Cleaning up, file renaming
